Remove the old osm2streets-based importer
