+++ examples/import.html
@@ -74,7 +74,7 @@
     <script type="module">
       import init, {
         convert,
-      } from "./osm-to-route-snapper-v2/pkg/osm_to_route_snapper_v2.js";
+      } from "./osm-to-route-snapper/pkg/osm_to_route_snapper.js";
 
       await init();
