--- examples/osm-to-route-snapper-v2
@@ -1,1 +1,0 @@
-../osm-to-route-snapper-v2
