--- osm-to-route-snapper-v2/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "osm-to-route-snapper-v2"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-anyhow = "1.0.75"
-bincode = "1.3.3"
-geom = { git = "https://github.com/a-b-street/abstreet" }
-log = "0.4.20"
-osm-reader = { git = "https://github.com/a-b-street/osm-reader" }
-route-snapper-graph = { path = "../route-snapper-graph" }
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-clap = { version = "4.4.6", features = ["derive"] }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-console_error_panic_hook = "0.1.6"
-console_log = "1.0.0"
-wasm-bindgen = "0.2.87"
-web-sys = { version = "0.3.64", features = ["console"] }
