+++ osm-to-route-snapper/Cargo.toml
@@ -7,12 +7,12 @@ edition = "2021"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-abstutil = { git = "https://github.com/a-b-street/abstreet" }
-bincode = "1.3.1"
+anyhow = "1.0.75"
+bincode = "1.3.3"
 geom = { git = "https://github.com/a-b-street/abstreet" }
-osm2streets = { git = "https://github.com/a-b-street/osm2streets" }
+log = "0.4.20"
+osm-reader = { git = "https://github.com/a-b-street/osm-reader" }
 route-snapper-graph = { path = "../route-snapper-graph" }
-streets_reader = { git = "https://github.com/a-b-street/osm2streets" }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 clap = { version = "4.4.6", features = ["derive"] }
@@ -20,9 +20,5 @@ clap = { version = "4.4.6", features = ["derive"] }
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 console_error_panic_hook = "0.1.6"
 console_log = "1.0.0"
-# TODO Upstream this in abstutil crate. WASM is missing some runtime dep otherwise.
-instant = { version = "0.1.12", features = ["wasm-bindgen"] }
-log = "0.4.20"
-serde-wasm-bindgen = "0.6.0"
 wasm-bindgen = "0.2.87"
 web-sys = { version = "0.3.64", features = ["console"] }
