+++ user_guide.md
@@ -7,14 +7,12 @@ The plugin can draw routes on any
 that has coordinates defined for the edges.
 
 A common use case is routing along a street network. You can create an example
-file from OpenStreetMap data using
-[osm2streets](https://github.com/a-b-street/osm2streets). The easiest way to do
-this for smaller areas is [in your web
-browser](https://dabreegster.github.io/route_snapper/import.html).
+file from OpenStreetMap data. The easiest way to do this for smaller areas is
+[in your web browser](https://dabreegster.github.io/route_snapper/import.html).
 
-For larger areas, you need an `.osm.xml` file, and optionally a GeoJSON file
-with one polygon representing the boundary of your area. You'll need to
-[install Rust](https://www.rust-lang.org/tools/install) to run this:
+For larger areas, you need an `.osm.xml` or `.osm.pbf` file, and optionally a
+GeoJSON file with one polygon representing the boundary of your area. You'll
+need to [install Rust](https://www.rust-lang.org/tools/install) to run this:
 
 ```
 cd osm-to-route-snapper
@@ -178,13 +176,13 @@ document.getElementById("snap-tool").addEventListener("new-route", () => {
 ## Routing caveats
 
 The routes calculated by the tool are based on the input graph. The default
-option described above uses osm2streets and pulls in road segments from
-OpenStreetMap for many modes, including tram or light-rail, walking or cycling
-only paths, and `highway=construction`. The "optimal" paths drawn by the tool
-are based on Euclidean distance -- no speed limits, safety of following the
-route by some user, etc is attempted. The route may violate one-way
-restrictions. In other words, if you're using the defaults, you will get routes
-that shouldn't actually be followed in the real world for many reasons.
+option described above pulls in road segments from OpenStreetMap for many
+modes, including tram or light-rail, walking or cycling only paths, and
+`highway=construction`. The "optimal" paths drawn by the tool are based on
+Euclidean distance -- no speed limits, safety of following the route by some
+user, etc is attempted. The route may violate one-way restrictions. In other
+words, if you're using the defaults, you will get routes that shouldn't
+actually be followed in the real world for many reasons.
 
 This default is designed for one particular use case: drawing potential new
 active travel routes along existing roads. The user designing these proposed
